backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlmodel==0.0.8
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==1.10.12
alembic==1.12.1
python-dotenv==1.0.0
